back-end/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
click==8.3.3
cryptography==47.0.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
iniconfig==2.3.0
Mako==1.3.12
MarkupSafe==3.0.3
packaging==26.2
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-decouple==3.8
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
